% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GINR.psat-package.R
\docType{package}
\name{GINR.psat-package}
\alias{GINR.psat}
\alias{GINR.psat-package}
\title{GINR.psat: Using and Visualizing Archival Tag data}
\description{
GINR.psat allows to easily visualize and analize data from archival tags. The package was created with WILDLIFE computers Inc. users in mind at the Greenland Institute of Natural Resources.
}
\author{
\strong{Maintainer}: Daniel Estevez \email{daeb@natur.gl} (\href{https://orcid.org/0000-0002-5852-8145}{ORCID})

}
\keyword{internal}
