% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tag.R
\name{read_tag}
\alias{read_tag}
\title{Read either argos or internal data from tags}
\usage{
read_tag(tagfile, type, temporal = NULL, tz = "GMT", delim = ",", keep = FALSE)
}
\arguments{
\item{tagfile}{string Path to the file where tag data is stored.}

\item{type}{string One of two values: "argos" or "internal".}

\item{temporal}{double Variable containing the time at which an individual observation took place. This variable is intended to make the internal variable of time readable in POSIXct format. Default NULL.}

\item{tz}{string Time zone to be used for modifying temporal variable in internal data. Set to UTC by default.}

\item{delim}{string The separation between columns in the data set. Either ",", " ", "\\t".}

\item{keep}{logical Specifies whether to keep the temporal variable which is tranformed to POSIXct. Default FALSE}
}
\value{
Object of class data.frame, tbl_df and tbl (same as in dplyr function read_delim)
}
\description{
read_tag loads a file provided by a Psat tag from Wildlife computers Inc and makes a number of adjustments from the existing variables on the data. The function works differently whether you got data from the satellite (argos) or from the tag itself (internal).
}
