% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_tag.R
\name{snap_tag}
\alias{snap_tag}
\title{Take a snap of the temporal data from internal tag data}
\usage{
snap_tag(
  tag,
  temporal,
  time_begin,
  time_end,
  temperature = TRUE,
  depth = TRUE,
  light_level = TRUE,
  interval,
  depth_range,
  tmp_range = NULL,
  main = NULL
)
}
\arguments{
\item{tag}{tbl or data.frame Object containing the tag information}

\item{temporal}{string Name of the variable containing the date and time}

\item{time_begin}{string Starting point of the time window}

\item{time_end}{string Ending point of the time window}

\item{temperature}{logical Set whether to present temperature in the plot}

\item{depth}{logical Set whether to present depth in the plot}

\item{light_level}{logical Set whether to present the light level in the plot}

\item{interval}{string indicating by how much time should date ticks appear (e.g. "12 hour" or "5 min")}

\item{depth_range}{vector indicating desired Depth limits to be ploted}

\item{tmp_range}{vector indicating desired Temperature limits to be ploted}

\item{main}{string, title to be included in the plot, default NULL}
}
\value{
Base R plot
}
\description{
This function produces a plot for a time range from the tag observations. The plot shows depth, light level and temperature.
}
